+++ src/main/java/org/opentripplanner/graph_builder/GraphBuilder.java
@@ -173,6 +173,7 @@ public class GraphBuilder implements Runnable {
             LOG.error("'{}' is not a readable directory.", dir);
             return null;
         }
+
         // Find and parse config files first to reveal syntax errors early without waiting for graph build.
         builderConfig = OTPMain.loadJson(new File(dir, BUILDER_CONFIG_FILENAME));
         GraphBuilderParameters builderParams = new GraphBuilderParameters(builderConfig);
@@ -266,30 +267,36 @@ public class GraphBuilder implements Runnable {
         graphBuilder.addModule(streetLinkerModule);
         // Load elevation data and apply it to the streets.
         // We want to do run this module after loading the OSM street network but before finding transfers.
+        ElevationGridCoverageFactory gcf = null;
         if (builderParams.elevationBucket != null) {
             // Download the elevation tiles from an Amazon S3 bucket
             S3BucketConfig bucketConfig = builderParams.elevationBucket;
             File cacheDirectory = new File(params.cacheDirectory, "ned");
             DegreeGridNEDTileSource awsTileSource = new DegreeGridNEDTileSource();
-            awsTileSource = new DegreeGridNEDTileSource();
             awsTileSource.awsAccessKey = bucketConfig.accessKey;
             awsTileSource.awsSecretKey = bucketConfig.secretKey;
             awsTileSource.awsBucketName = bucketConfig.bucketName;
-            NEDGridCoverageFactoryImpl gcf = new NEDGridCoverageFactoryImpl(cacheDirectory);
-            gcf.tileSource = awsTileSource;
-            GraphBuilderModule elevationBuilder = new ElevationModule(gcf, builderParams.elevationUnitMultiplier);
-            graphBuilder.addModule(elevationBuilder);
+            gcf = new NEDGridCoverageFactoryImpl(cacheDirectory, awsTileSource);
         } else if (builderParams.fetchElevationUS) {
             // Download the elevation tiles from the official web service
             File cacheDirectory = new File(params.cacheDirectory, "ned");
-            ElevationGridCoverageFactory gcf = new NEDGridCoverageFactoryImpl(cacheDirectory);
-            GraphBuilderModule elevationBuilder = new ElevationModule(gcf, builderParams.elevationUnitMultiplier);
-            graphBuilder.addModule(elevationBuilder);
+            gcf = new NEDGridCoverageFactoryImpl(cacheDirectory);
         } else if (demFile != null) {
             // Load the elevation from a file in the graph inputs directory
-            ElevationGridCoverageFactory gcf = new GeotiffGridCoverageFactoryImpl(demFile);
-            GraphBuilderModule elevationBuilder = new ElevationModule(gcf, builderParams.elevationUnitMultiplier);
-            graphBuilder.addModule(elevationBuilder);
+            gcf = new GeotiffGridCoverageFactoryImpl(demFile);
+        }
+        if (gcf != null) {
+            graphBuilder.addModule(
+                new ElevationModule(
+                    gcf,
+                    params.cacheDirectory,
+                    builderParams.readCachedElevations,
+                    builderParams.writeCachedElevations,
+                    builderParams.elevationUnitMultiplier,
+                    builderParams.includeEllipsoidToGeoidDifference,
+                    builderParams.multiThreadElevationCalculations
+                )
+            );
         }
         if ( hasGTFS ) {
             // The stops can be linked to each other once they are already linked to the street network.