+++ docs/Changelog.md
@@ -7,6 +7,7 @@
 - Fixes surefire test failure during build (#2816)
 - Improve documentation for `mode` routing parameter (#2809)
 - Disable linking from already linked stops (#2372)
+- Optimize elevation calculations
 
 ## 1.4 (2019-07-30)